fix: loading family cards
